Migrate `check_normalized` and `set_not_normalized` to `general.py`
